Create GF256PLY.z80
Mostly worked out in Mimas and then run through 8xvtoasm, so the formatting is a little screwy, but I'll fix it later. DO NOT RUN THIS CODE ON AN ACTUAL CALCULATOR. I haven't confirmed whether it's 100% safe yet. (I mean, as-is, it *should* be fine; it doesn't crash my 84+. But still, part of the debug code relies on some 84+-specific ports to slow the thing down.) There is next to zero error handling here: Garbage in, bricked machine out.